// src/store/app/types.ts
export enum appTypes {
  USER_REGISTRATION_START = 'APP/USER_REGISTRATION_START',
  USER_LOGIN_START = 'USER_LOGIN_START',
  USER_LOGIN_SUCCESS = 'USER_LOGIN_SUCCESS',
  FETCH_SUCCESS = 'APP/FETCH_SUCCESS',
  FETCH_PROFILE_START = 'FETCH_PROFILE_START',
  FETCH_PROFILE_SUCCESS = 'FETCH_PROFILE_SUCCESS',
  FETCH_TASKS_START = 'FETCH_TASKS_START',
  FETCH_TASKS_SUCCESS = 'FETCH_TASKS_SUCCESS',
  FETCH_ERROR_SHOW = 'APP/FETCH_ERROR_SHOW',
  FETCH_ERROR_HIDE = 'APP/FETCH_ERROR_HIDE',
  REFRESH_TOKEN_START = 'REFRESH_TOKEN_START',
  REFRESH_TOKEN_UPDATED = 'REFRESH_TOKEN_UPDATED',
  TOGGLE_CREATE_TASK_MODAL = 'MODAL/TOGGLE_CREATE_TASK_MODAL',
  // Tasks creating
  CREATE_TASK_START = 'TASK/TASK_CREATE_START',
  CREATE_TASK_SUCCESS = 'TASK/CREATE_TASK_SUCCESS',
}